[StaffNotes] Move controller and views
