--- app/views/admin/staff_notes/partials/_list_of_notes.html.erb
@@ -1,3 +1,0 @@
-<div id="p-staff-notes-list" class="staff-note-list">
-  <%= render partial: "admin/staff_notes/partials/staff_note", collection: staff_notes, locals: { show_receiver_name: show_receiver_name } %>
-</div>
