+++ app/views/staff_notes/edit.html.erb
@@ -2,7 +2,7 @@
   <div id="a-edit">
     <h1>Edit Staff Note For <%= link_to_user @staff_note.user %></h1>
 
-    <%= render "admin/staff_notes/partials/edit", staff_note: @staff_note %>
+    <%= render "staff_notes/partials/edit", staff_note: @staff_note %>
   </div>
 </div>
