+++ app/views/staff_notes/index.html.erb
@@ -7,7 +7,7 @@
     <% end %>
     <%= render "search" %>
 
-    <%= render "admin/staff_notes/partials/list_of_notes", staff_notes: @notes, show_receiver_name: @user.blank? %>
+    <%= render "staff_notes/partials/list_of_notes", staff_notes: @notes, show_receiver_name: @user.blank? %>
     <%= numbered_paginator(@notes) %>
   </div>
 </div>
