+++ app/views/staff_notes/new.html.erb
@@ -2,7 +2,7 @@
   <div id="a-new">
     <h1>New Staff Note for <%= link_to_user(@user) %></h1>
 
-    <%= render "admin/staff_notes/partials/new", staff_note: @staff_note, user: @user %>
+    <%= render "staff_notes/partials/new", staff_note: @staff_note, user: @user %>
   </div>
 </div>
