+++ app/views/staff_notes/partials/_for_user.html.erb
@@ -4,10 +4,10 @@
       <summary>Staff Notes (<%= user.staff_notes.count %>)</summary>
       <div>
         <h4><%= link_to "Staff Notes", staff_notes_path(user_id: user.id) %></h4>
-        <%= render "admin/staff_notes/partials/list_of_notes", staff_notes: user.staff_notes.limit(15), show_receiver_name: false %>
+        <%= render "staff_notes/partials/list_of_notes", staff_notes: user.staff_notes.limit(15), show_receiver_name: false %>
         <div class="new-staff-note">
           <p><%= link_to "Create »", new_staff_note_path(user_id: user.id), class: "expand-new-staff-note" %></p>
-          <%= render "admin/staff_notes/partials/new", user: user, staff_note: StaffNote.new(user_id: user.id), hidden: true %>
+          <%= render "staff_notes/partials/new", user: user, staff_note: StaffNote.new(user_id: user.id), hidden: true %>
         </div>
       </div>
     </details>
