+++ app/views/staff_notes/show.html.erb
@@ -2,7 +2,7 @@
   <div id="a-show">
     <h1>Staff Note For <%= link_to_user @staff_note.user %></h1>
 
-    <%= render "admin/staff_notes/partials/staff_note", staff_note: @staff_note, show_receiver_name: true %>
+    <%= render "staff_notes/partials/staff_note", staff_note: @staff_note, show_receiver_name: true %>
   </div>
 </div>
