+++ app/views/static/site_map.html.erb
@@ -119,6 +119,9 @@
           <li><%= link_to("Stuck DNP tags", new_admin_stuck_dnp_path) %></li>
           <li><%= link_to("Destroyed Posts", admin_destroyed_posts_path) %></li>
         <% end %>
+        <% if CurrentUser.is_staff? %>
+          <li><%= link_to("Staff Notes", staff_notes_path) %></li>
+        <% end %>
         <li><%= link_to("Upload Whitelist", upload_whitelists_path) %></li>
         <li><%= link_to("Mod Actions", mod_actions_path) %></li>
         <li><%= link_to("Bulk Update Requests", bulk_update_requests_path) %></li>
